Moved methods to proper dao classes
IBP-3212
