+++ src/main/java/org/generationcp/middleware/dao/dms/DmsProjectDao.java
@@ -1385,85 +1385,4 @@ public class DmsProjectDao extends GenericDAO<DmsProject, Integer> {
 		}
 	}
 
-	public List<MeasurementVariable> getEnvironmentConditionVariablesByGeoLocationIdAndVariableIds(Integer geolocationId, List<Integer> variableIds) {
-		List<MeasurementVariable> studyVariables = new ArrayList<>();
-
-		try{
-			final SQLQuery query =
-				this.getSession().createSQLQuery("SELECT envcvt.name AS name, envcvt.definition AS definition, "
-					+ "		cvt_scale.name AS scaleName, pheno.value AS value from phenotype pheno "
-					+ "		INNER JOIN cvterm envcvt ON envcvt.cvterm_id = pheno.observable_id AND envcvt.cvterm_id IN (:variableIds) "
-					+ "		INNER JOIN cvterm_relationship cvt_rel ON cvt_rel.subject_id = envcvt.cvterm_id AND cvt_rel.type_id = " + TermId.HAS_SCALE.getId()
-					+ "     INNER JOIN cvterm cvt_scale ON cvt_scale.cvterm_id = cvt_rel.object_id\n"
-					+ "     INNER JOIN nd_experiment envnde ON  pheno.nd_experiment_id = envnde.nd_experiment_id\n"
-					+ "		INNER JOIN nd_geolocation gl ON envnde.nd_geolocation_id = gl.nd_geolocation_id AND gl.nd_geolocation_id = :geolocationId ;");
-			query.addScalar("name", new StringType());
-			query.addScalar("definition", new StringType());
-			query.addScalar("scaleName", new StringType());
-			query.addScalar("value", new StringType());
-			query.setParameterList("variableIds", variableIds);
-			query.setParameter("geolocationId", geolocationId);
-
-			final List<Object> results = query.list();
-			for(Object result: results) {
-				final Object[] row = (Object[]) result;
-				final MeasurementVariable measurementVariable = new MeasurementVariable();
-				measurementVariable.setName((row[0] instanceof String) ? (String) row[0] : null);
-				measurementVariable.setDescription((row[1] instanceof String) ? (String) row[1] : null);
-				measurementVariable.setScale((row[2] instanceof String) ? (String) row[2] : null);
-				measurementVariable.setValue((row[3] instanceof String) ? (String) row[3] : null);
-				studyVariables.add(measurementVariable);
-			}
-		} catch (final MiddlewareQueryException e) {
-			final String message = "Error with getEnvironmentConditionVariablesByGeoLocationIdAndVariableIds() query from geolocationId: " + geolocationId
-				+ " and variableIds: " + variableIds;
-			DmsProjectDao.LOG.error(message, e);
-			throw new MiddlewareQueryException(message, e);
-		}
-		return studyVariables;
-	}
-
-	public List<MeasurementVariable> getEnvironmentDetailVariablesByGeoLocationIdAndVariableIds(Integer geolocationId, List<Integer> variableIds) {
-		List<MeasurementVariable> studyVariables = new ArrayList<>();
-		final List<Integer> standardEnvironmentFactors = Lists.newArrayList(
-			TermId.LOCATION_ID.getId(),
-			TermId.TRIAL_INSTANCE_FACTOR.getId(),
-			TermId.EXPERIMENT_DESIGN_FACTOR.getId());
-		try{
-			final SQLQuery query =
-				this.getSession().createSQLQuery("SELECT ispcvt.name as name, ispcvt.definition as definition, "
-					+ "		cvt_scale.name AS scaleName, gprop.value AS value FROM nd_geolocationprop gprop "
-					+ "		INNER JOIN cvterm ispcvt ON ispcvt.cvterm_id = gprop.type_id AND ispcvt.cvterm_id in (:variableIds) "
-					+ "		INNER JOIN cvterm_relationship cvt_rel ON cvt_rel.subject_id = ispcvt.cvterm_id AND cvt_rel.type_id = " + TermId.HAS_SCALE.getId()
-					+ "		INNER JOIN cvterm cvt_scale ON cvt_scale.cvterm_id = cvt_rel.object_id "
-					+ "		INNER JOIN nd_geolocation gl ON gprop.nd_geolocation_id = gl.nd_geolocation_id "
-					+ "	    WHERE gl.nd_geolocation_id = :geolocationId AND ispcvt.cvterm_id NOT IN (:standardEnvironmentFactors) ;");
-			query.addScalar("name", new StringType());
-			query.addScalar("definition", new StringType());
-			query.addScalar("scaleName", new StringType());
-			query.addScalar("value", new StringType());
-			query.setParameterList("variableIds", variableIds);
-			query.setParameter("geolocationId", geolocationId);
-			query.setParameterList("standardEnvironmentFactors", standardEnvironmentFactors);
-
-			final List<Object> results = query.list();
-			for(Object result: results) {
-
-				final Object[] row = (Object[]) result;
-				final MeasurementVariable measurementVariable = new MeasurementVariable();
-				measurementVariable.setName((row[0] instanceof String) ? (String) row[0] : null);
-				measurementVariable.setDescription((row[1] instanceof String) ? (String) row[1] : null);
-				measurementVariable.setScale((row[2] instanceof String) ? (String) row[2] : null);
-				measurementVariable.setValue((row[3] instanceof String) ? (String) row[3] : null);
-				studyVariables.add(measurementVariable);
-			}
-		} catch (final MiddlewareQueryException e) {
-			final String message = "Error with getEnvironmentConditionVariablesByGeoLocationIdAndVariableIds() query from geolocationId: " + geolocationId
-				+ " and variableIds: " + variableIds;
-			DmsProjectDao.LOG.error(message, e);
-			throw new MiddlewareQueryException(message, e);
-		}
-		return studyVariables;
-	}
-
 }
